+++ src/worker/util/changes.basketball.ts
@@ -305,6 +305,11 @@ const changes = [
 		msg:
 			'You can customize the order and visibility of columns in most tables by pressing the <span class="glyphicon glyphicon-option-vertical"></span> button to the top right of the table and clicking "Customize Columns". Also, my apologies for changing the default positioning of PTS on some tables! I returned it to its previous default position.',
 	},
+	{
+		date: "2020-07-01",
+		msg:
+			"The Power Rankings page now shows team rankings for individual rating categories.",
+	},
 ];
 
 export default changes;