+++ CHANGELOG.md
@@ -2,7 +2,7 @@ This is a user-focused changelog. The entries here are things that affect gamepl
 
 Entries in bold are ones important enough that I would have put them in [the old changelog](./CHANGELOG-old.md).
 
-* DEV: The Power Rankings page now shows more granular ratings broken down by position (in FBGM) or rating (in BBGM). [[1]]()
+* 2020-07-01: The Power Rankings page now shows team rankings broken down by position (in FBGM) or rating (in BBGM). [[1]](https://old.reddit.com/r/BasketballGM/comments/hjj3d5/the_power_rankings_page_now_shows_team_rankings/) [[2]](https://old.reddit.com/r/Football_GM/comments/hjj8en/the_power_rankings_page_now_shows_team_rankings/)
 
 * 2020-06-29: Logos for all the new teams! [[1]](https://basketball-gm.com/blog/2020/06/logos-for-the-new-teams/) [[2]](https://old.reddit.com/r/BasketballGM/comments/hid1tg/logos_for_all_the_new_teams/)